refactor(venue_nest): setup environment variables and logger same core service
