+++ database/constants.py
@@ -1,3 +1,3 @@
-from env_config import get_env_variable
+from core.service import get_env_variable
 
 DATABASE_URL = get_env_variable("DATABASE_URL")
