+++ database/service.py
@@ -1,5 +1,5 @@
-from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
 from sqlalchemy import create_engine
 from .constants import DATABASE_URL
